refactor(core): move maskToRegExpMask.js to utils, fix JSDoc typos
