+++ src/filter.js
@@ -1,6 +1,6 @@
 // eslint-disable-next-line import/no-extraneous-dependencies
 import conformToMask from 'text-mask-core/src/conformToMask';
-import { stringMaskToRegExpMask } from './maskToRegExpMask';
+import { stringMaskToRegExpMask } from './utils/maskToRegExpMask';
 import { isString } from './utils';
 
 /**
